Added checks and meaningful error messages for missing ID, ParentEntityID fields. Added unit tests.

+++ src/main/java/com/bullhorn/dataloader/task/DeleteAttachmentTask.java
@@ -15,6 +15,7 @@ import com.bullhornsdk.data.model.entity.core.type.BullhornEntity;
 import com.bullhornsdk.data.model.entity.core.type.FileEntity;
 import com.bullhornsdk.data.model.response.file.FileApiResponse;
 
+import java.io.IOException;
 import java.util.Map;
 
 /**
@@ -54,8 +55,13 @@ public class DeleteAttachmentTask<A extends AssociationEntity, E extends EntityA
         return Result.Delete(fileApiResponse.getFileId());
     }
 
-    private <F extends FileEntity> FileApiResponse deleteFile() {
+    private <F extends FileEntity> FileApiResponse deleteFile() throws IOException {
+        if (!dataMap.containsKey(StringConsts.PARENT_ENTITY_ID) || dataMap.get(StringConsts.PARENT_ENTITY_ID).isEmpty()) {
+            throw new IOException("Row " + rowNumber + ": Missing the '" + StringConsts.PARENT_ENTITY_ID + "' column required for deleteAttachments");
+        }
+        if (!dataMap.containsKey(StringConsts.ID) || dataMap.get(StringConsts.ID).isEmpty()) {
+            throw new IOException("Row " + rowNumber + ": Missing the '" + StringConsts.ID + "' column required for deleteAttachments");
+        }
         return bullhornData.deleteFile((Class<F>) entityClass, Integer.valueOf(dataMap.get(StringConsts.PARENT_ENTITY_ID)), Integer.valueOf(dataMap.get(StringConsts.ID)));
     }
-
 }
